finished the CRUD with blogs

+++ backend/app/Http/Controllers/API/BlogController.php
@@ -15,8 +15,12 @@ class BlogController extends BaseController
     /**
      * Read Multi
      */
-    public function index()
+    public function index(Request $request)
     {
+        $page = isset($request->page) ? $request->page : 1;
+        $countPerPage = isset($request->countPerPage) ? $request->countPerPage : 10;
+        $search = isset($request->search) ? $request->search : "";
+        $orderBy = isset($request->orderBy) ? $request->orderBy : "updated_at";
         $blogs = Blog::all();
         return $this->sendResponse(BlogResource::collection($blogs), 'Posts fetched.');
     }
@@ -28,8 +32,8 @@ class BlogController extends BaseController
     {
         $input = $request->all();
         $validator = Validator::make($input, [
-            'title' => 'required',
-            'description' => 'required'
+            'title' => 'required|min:10',
+            'description' => 'required|min:10'
         ]);
         if($validator->fails()){
             return $this->sendError($validator->errors());       
@@ -81,8 +85,12 @@ class BlogController extends BaseController
     /**
      * Delete
      */
-    public function destroy(Blog $blog)
+    public function destroy($id)
     {
+        $blog = Blog::find($id);
+        if (is_null($blog)) {
+            return $this->sendError('Post does not exist.');
+        }
         $blog->delete();
         return $this->sendResponse([], 'Post deleted.');
     }
